# backend/app/admin/service/cmd/server/assets/openapi.yaml
...
openapi: 3.0.3
info:
    title: Kratos巨石应用实践API
    description: Kratos巨石应用实践API
    contact:
        name: tx7do
        url: https://github.com/tx7do/kratos-monolithic-demo
        email: yanglinbo@gmail.com
    license:
        name: MIT License
        url: https://github.com/tx7do/kratos-monolithic-demo/blob/master/LICENSE
    version: "1.0"
paths:
...
    /admin/v1/orgs:
        get:
            tags:
                - OrganizationService
            description: 查询部门列表
            operationId: OrganizationService_ListOrganization
            parameters:
                - name: page
                  in: query
                  description: 当前页码
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  description: 每页的行数
                  schema:
                    type: integer
                    format: int32
                - name: query
                  in: query
                  description: 与过滤参数
                  schema:
                    type: string
                - name: or
                  in: query
                  description: 或过滤参数
                  schema:
                    type: string
                - name: orderBy
                  in: query
                  description: 排序条件
                  schema:
                    type: array
                    items:
                        type: string
                - name: nopaging
                  in: query
                  description: 是否不分页
                  schema:
                    type: boolean
                - name: field_mask
                  in: query
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListOrganizationResponse'
            security:
                - OAuth2PasswordBearer: []
        post:
            tags:
                - OrganizationService
            description: 创建部门
            operationId: OrganizationService_CreateOrganization
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateOrganizationRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Organization'
            security:
                - OAuth2PasswordBearer: []
    /admin/v1/orgs/{id}:
        get:
            tags:
                - OrganizationService
            description: 查询部门详情
            operationId: OrganizationService_GetOrganization
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Organization'
            security:
                - OAuth2PasswordBearer: []
        put:
            tags:
                - OrganizationService
            description: 更新部门
            operationId: OrganizationService_UpdateOrganization
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
                - name: operatorId
                  in: query
                  schema:
                    type: integer
                    format: uint32
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Organization'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Organization'
            security:
                - OAuth2PasswordBearer: []
        delete:
            tags:
                - OrganizationService
            description: 删除部门
            operationId: OrganizationService_DeleteOrganization
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
                - name: operatorId
                  in: query
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content: {}
            security:
                - OAuth2PasswordBearer: []
...
components:
    schemas:
...
        CreateOrganizationRequest:
            type: object
            properties:
                org:
                    $ref: '#/components/schemas/Organization'
                operatorId:
                    type: integer
                    format: uint32
            description: 创建部门 - 请求
...
        ListOrganizationResponse:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/Organization'
                total:
                    type: integer
                    format: int32
            description: 部门列表 - 答复
...
        Organization:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                name:
                    type: string
                parentId:
                    type: integer
                    format: uint32
                orderNo:
                    type: integer
                    format: int32
                status:
                    enum:
                        - ON
                        - OFF
                    type: string
                    default: ON
                    description: 状态
                remark:
                    type: string
                children:
                    type: array
                    items:
                        $ref: '#/components/schemas/Organization'
                createTime:
                    type: string
                updateTime:
                    type: string
                deleteTime:
                    type: string
            description: 部门
...
    securitySchemes:
        OAuth2PasswordBearer:
            type: oauth2
            flows:
                password:
                    tokenUrl: /admin/v1/login
                    refreshUrl: /admin/v1/refresh_token
                    scopes: {}
...